Add dynamic data changes to list view converter sample
The converters for the ListView do not automatically get invoked when one
of the list cells is changed, which is not what I expected. I needed to
play around with a few things in order to figure out how to do it.
I found a StackOverflow answer that said that conversions will only run like
this if a MultiBinding is used.
https://stackoverflow.com/questions/8888480/why-is-my-converter-based-on-listview-itemssource-not-being-called

I still am not sure why it needs to be a MultiBinding and why the source of
the ListView needs to be one of the inputs. However, it seems to be working
as I want it to right now. The code could use some clean-up as well.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>