Perform database migration and call Migrate()
I had a difficult time figuring out how to get the db.Database.Migrate()
call to not throw an exception. The trick is to install the SQLit package
in the WPF executable project (i.e. the View) even though the View doesn't
ever use the DbContext (the ViewModel calls Migrate()).

So the schema is defined and the database is migrated (which creates it if it
does not already exist) but the app doesn't do anything with it at this time.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>