Add FoodItemView to show details of selected item in Nutrition sample
Created a FoodItem_VM to store the FoodItem. It is used by FoodItemView,
which is a UserControl that is placed in the nutrition sample. It only
displays a few of the attributes of the food item, just enough to be a
useful sample.

Also added EFCore to the WpfApp_View project, because I was having the
problem with e_sqlite3.dll disappearing again.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>